+++ frost-ristretto255/src/frost.rs
@@ -21,9 +21,10 @@
 //! Internally, keygen_with_dealer generates keys using Verifiable Secret
 //! Sharing, where shares are generated using Shamir Secret Sharing.
 
-use std::{convert::TryFrom, fmt::Debug};
+use std::{collections::HashMap, convert::TryFrom, fmt, fmt::Debug};
 
-use curve25519_dalek::scalar::Scalar;
+use curve25519_dalek::{ristretto::RistrettoPoint, scalar::Scalar, traits::Identity};
+
 use hex::FromHex;
 
 pub mod keys;
@@ -33,7 +34,7 @@ pub mod round2;
 #[cfg(test)]
 mod tests;
 
-use crate::{generate_challenge, Signature, H1};
+use crate::{generate_challenge, Signature, H1, H3};
 
 /// The binding factor, also known as _rho_ (ρ)
 ///
@@ -44,8 +45,8 @@ use crate::{generate_challenge, Signature, H1};
 #[derive(Clone, Debug, PartialEq)]
 struct Rho(Scalar);
 
-impl From<&round2::SigningPackage> for Rho {
-    fn from(signing_package: &round2::SigningPackage) -> Rho {
+impl From<&SigningPackage> for Rho {
+    fn from(signing_package: &SigningPackage) -> Rho {
         let preimage = signing_package.rho_preimage();
 
         let binding_factor = H1(&preimage[..]);
@@ -81,7 +82,7 @@ impl TryFrom<[u8; 32]> for Rho {
 /// Generates the lagrange coefficient for the i'th participant.
 fn generate_lagrange_coeff(
     signer_index: u16,
-    signing_package: &round2::SigningPackage,
+    signing_package: &SigningPackage,
 ) -> Result<Scalar, &'static str> {
     let mut num = Scalar::one();
     let mut den = Scalar::one();
@@ -107,6 +108,111 @@ fn generate_lagrange_coeff(
     Ok(lagrange_coeff)
 }
 
+/// Generated by the coordinator of the signing operation and distributed to
+/// each signing party
+#[derive(Debug)]
+pub struct SigningPackage {
+    /// The set of commitments participants published in the first round of the
+    /// protocol.
+    signing_commitments: HashMap<u16, round1::SigningCommitments>,
+    /// Message which each participant will sign.
+    ///
+    /// Each signer should perform protocol-specific verification on the
+    /// message.
+    message: Vec<u8>,
+}
+
+impl SigningPackage {
+    /// Create a new `SigingPackage`
+    ///
+    /// The `signing_commitments` are sorted by participant `index`.
+    pub fn new(
+        mut signing_commitments: Vec<round1::SigningCommitments>,
+        message: Vec<u8>,
+    ) -> SigningPackage {
+        signing_commitments.sort_by_key(|a| a.index);
+
+        SigningPackage {
+            signing_commitments: signing_commitments
+                .into_iter()
+                .map(|s| (s.index, s))
+                .collect(),
+            message,
+        }
+    }
+
+    /// Get a signing commitment by its participant index.
+    pub fn signing_commitment(&self, index: &u16) -> round1::SigningCommitments {
+        self.signing_commitments[index]
+    }
+
+    /// Get the signing commitments, sorted by the participant indices
+    pub fn signing_commitments(&self) -> Vec<round1::SigningCommitments> {
+        let mut signing_commitments: Vec<round1::SigningCommitments> =
+            self.signing_commitments.values().cloned().collect();
+        signing_commitments.sort_by_key(|a| a.index);
+        signing_commitments
+    }
+
+    /// Get the message to be signed
+    pub fn message(&self) -> &Vec<u8> {
+        &self.message
+    }
+
+    /// Compute the preimage to H3 to compute rho
+    // We separate this out into its own method so it can be tested
+    pub(super) fn rho_preimage(&self) -> Vec<u8> {
+        let mut preimage = vec![];
+
+        preimage
+            .extend_from_slice(&round1::encode_group_commitments(self.signing_commitments())[..]);
+        preimage.extend_from_slice(&H3(self.message.as_slice()));
+
+        preimage
+    }
+}
+
+/// The product of all signers' individual commitments, published as part of the
+/// final signature.
+#[derive(PartialEq)]
+pub struct GroupCommitment(pub(super) RistrettoPoint);
+
+impl Debug for GroupCommitment {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.debug_tuple("GroupCommitment")
+            .field(&hex::encode(self.0.compress().to_bytes()))
+            .finish()
+    }
+}
+
+impl TryFrom<&SigningPackage> for GroupCommitment {
+    type Error = &'static str;
+
+    /// Generates the group commitment which is published as part of the joint
+    /// Schnorr signature.
+    fn try_from(signing_package: &SigningPackage) -> Result<GroupCommitment, &'static str> {
+        let rho: Rho = signing_package.into();
+
+        let identity = RistrettoPoint::identity();
+        let mut accumulator = identity;
+
+        // Ala the sorting of B, just always sort by index in ascending order
+        //
+        // https://github.com/cfrg/draft-irtf-cfrg-frost/blob/master/draft-irtf-cfrg-frost.md#encoding-operations-dep-encoding
+        for commitment in signing_package.signing_commitments() {
+            // The following check prevents a party from accidentally revealing their share.
+            // Note that the '&&' operator would be sufficient.
+            if identity == commitment.binding.0 || identity == commitment.hiding.0 {
+                return Err("Commitment equals the identity.");
+            }
+
+            accumulator += commitment.hiding.0 + (commitment.binding.0 * rho.0)
+        }
+
+        Ok(GroupCommitment(accumulator))
+    }
+}
+
 ////////////////////////////////////////////////////////////////////////////////
 // Aggregation
 ////////////////////////////////////////////////////////////////////////////////
@@ -127,11 +233,11 @@ fn generate_lagrange_coeff(
 /// can avoid that step. However, at worst, this results in a denial of
 /// service attack due to publishing an invalid signature.
 pub fn aggregate(
-    signing_package: &round2::SigningPackage,
+    signing_package: &SigningPackage,
     signing_shares: &[round2::SignatureShare],
     pubkeys: &keys::PublicKeyPackage,
 ) -> Result<Signature, &'static str> {
-    let group_commitment = round2::GroupCommitment::try_from(signing_package)?;
+    let group_commitment = GroupCommitment::try_from(signing_package)?;
 
     let challenge = generate_challenge(
         &group_commitment.0.compress().to_bytes(),
