+++ frost-ristretto255/src/frost/round2.rs
@@ -1,86 +1,15 @@
 //! FROST Round 2 functionality and types, for signature share generation
 
-use std::{
-    collections::HashMap,
-    fmt::{self, Debug},
-};
+use std::fmt::{self, Debug};
 
-use curve25519_dalek::{
-    constants::RISTRETTO_BASEPOINT_POINT, ristretto::RistrettoPoint, scalar::Scalar,
-    traits::Identity,
-};
+use curve25519_dalek::{constants::RISTRETTO_BASEPOINT_POINT, scalar::Scalar};
 
 use zeroize::DefaultIsZeroes;
 
 use crate::{
-    frost::{self, round1},
-    generate_challenge, H3,
+    frost::{self, round1, *},
+    generate_challenge,
 };
-
-/// Generated by the coordinator of the signing operation and distributed to
-/// each signing party
-#[derive(Debug)]
-pub struct SigningPackage {
-    /// The set of commitments participants published in the first round of the
-    /// protocol.
-    signing_commitments: HashMap<u16, round1::SigningCommitments>,
-    /// Message which each participant will sign.
-    ///
-    /// Each signer should perform protocol-specific verification on the
-    /// message.
-    message: Vec<u8>,
-}
-
-impl SigningPackage {
-    /// Create a new `SigingPackage`
-    ///
-    /// The `signing_commitments` are sorted by participant `index`.
-    pub fn new(
-        mut signing_commitments: Vec<round1::SigningCommitments>,
-        message: Vec<u8>,
-    ) -> SigningPackage {
-        signing_commitments.sort_by_key(|a| a.index);
-
-        SigningPackage {
-            signing_commitments: signing_commitments
-                .into_iter()
-                .map(|s| (s.index, s))
-                .collect(),
-            message,
-        }
-    }
-
-    /// Get a signing commitment by its participant index.
-    pub fn signing_commitment(&self, index: &u16) -> round1::SigningCommitments {
-        self.signing_commitments[index]
-    }
-
-    /// Get the signing commitments, sorted by the participant indices
-    pub fn signing_commitments(&self) -> Vec<round1::SigningCommitments> {
-        let mut signing_commitments: Vec<round1::SigningCommitments> =
-            self.signing_commitments.values().cloned().collect();
-        signing_commitments.sort_by_key(|a| a.index);
-        signing_commitments
-    }
-
-    /// Get the message to be signed
-    pub fn message(&self) -> &Vec<u8> {
-        &self.message
-    }
-
-    /// Compute the preimage to H3 to compute rho
-    // We separate this out into its own method so it can be tested
-    pub(super) fn rho_preimage(&self) -> Vec<u8> {
-        let mut preimage = vec![];
-
-        preimage.extend_from_slice(
-            &round1::encode_group_commitments(self.signing_commitments())[..],
-        );
-        preimage.extend_from_slice(&H3(self.message.as_slice()));
-
-        preimage
-    }
-}
 
 /// A representation of a single signature share used in FROST structures and messages, including
 /// the group commitment share.
@@ -152,47 +81,6 @@ impl SignatureShare {
             return Err("Invalid signature share");
         }
         Ok(())
-    }
-}
-
-/// The product of all signers' individual commitments, published as part of the
-/// final signature.
-#[derive(PartialEq)]
-pub struct GroupCommitment(pub(super) RistrettoPoint);
-
-impl Debug for GroupCommitment {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_tuple("GroupCommitment")
-            .field(&hex::encode(self.0.compress().to_bytes()))
-            .finish()
-    }
-}
-
-impl TryFrom<&SigningPackage> for GroupCommitment {
-    type Error = &'static str;
-
-    /// Generates the group commitment which is published as part of the joint
-    /// Schnorr signature.
-    fn try_from(signing_package: &SigningPackage) -> Result<GroupCommitment, &'static str> {
-        let rho: frost::Rho = signing_package.into();
-
-        let identity = RistrettoPoint::identity();
-        let mut accumulator = identity;
-
-        // Ala the sorting of B, just always sort by index in ascending order
-        //
-        // https://github.com/cfrg/draft-irtf-cfrg-frost/blob/master/draft-irtf-cfrg-frost.md#encoding-operations-dep-encoding
-        for commitment in signing_package.signing_commitments() {
-            // The following check prevents a party from accidentally revealing their share.
-            // Note that the '&&' operator would be sufficient.
-            if identity == commitment.binding.0 || identity == commitment.hiding.0 {
-                return Err("Commitment equals the identity.");
-            }
-
-            accumulator += commitment.hiding.0 + (commitment.binding.0 * rho.0)
-        }
-
-        Ok(GroupCommitment(accumulator))
     }
 }
